refactor: ServiceImpl 삭제 리팩토링 - Book, BookReport 도메인에서 오버엔지니어링 요소였던, ServiceImpl을 제거하고 Service를 클래스로 바꾸고 병합했습니다.
