+++ src/main/java/com/kernel360/boogle/bookreport/service/BookReportService.java
@@ -1,25 +1,72 @@
 package com.kernel360.boogle.bookreport.service;
 
 import com.kernel360.boogle.bookreport.db.BookReportEntity;
+import com.kernel360.boogle.bookreport.db.BookReportRepository;
 import com.kernel360.boogle.bookreport.model.BookReportDTO;
 import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
+import java.time.LocalDateTime;
 import java.util.Optional;
 
-public interface BookReportService {
-    public void createBookReport(BookReportDTO bookReport);
+@Service
+@Transactional
+public class BookReportService {
 
-    Optional<BookReportEntity> getBookReportById(Long bookReportId);
+    private final BookReportRepository bookReportRepository;
 
-    Page<BookReportEntity> getPublicBookReports(String isPublic, int page);
+    private static final Integer PAGE_SIZE = 6;
 
-    Page<BookReportEntity> getMyBookReports(Long memberId, int page);
+    public BookReportService(BookReportRepository bookReportRepository) {
+        this.bookReportRepository = bookReportRepository;
+    }
 
-    Page<BookReportEntity> getAllBookReports(int page);
+    public void createBookReport(BookReportDTO bookReport) {
+        bookReport.getBookReportEntity().setCreatedBy("TEST"); // 로그인 사용자 정보 들어가야 함
+        bookReport.getBookReportEntity().setMemberId(1L); // 로그인 사용자 정보 들어가야 함
+        bookReportRepository.save(bookReport.getBookReportEntity());
+    }
 
-    public void updateBookReport(BookReportDTO bookReport);
+    public Optional<BookReportEntity> getBookReportById(Long bookReportId) {
+        return bookReportRepository.findById(bookReportId);
+    }
 
-    public void deleteBookReport(Long bookReportId);
+    public Page<BookReportEntity> getPublicBookReports(String isPublic, int page) {
+        Pageable pageable = PageRequest.of(page, PAGE_SIZE);
+        return bookReportRepository.findAllByIsPublicEqualsAndIsDeletedNotOrderByIdDesc(isPublic, "Y", pageable);
+    }
+
+    public Page<BookReportEntity> getMyBookReports(Long memberId, int page) {
+        Pageable pageable = PageRequest.of(page, PAGE_SIZE);
+        return bookReportRepository.findAllByMemberIdEqualsAndIsDeletedNotOrderByIdDesc(memberId, "Y", pageable);
+    }
+
+    public Page<BookReportEntity> getAllBookReports(int page) {
+        Pageable pageable = PageRequest.of(page, PAGE_SIZE);
+        return bookReportRepository.findAllByIsDeletedNotOrderByIdDesc("Y", pageable);
+    }
+
+    public void updateBookReport(BookReportDTO bookReport) {
+        bookReport.getBookReportEntity().setLastModifiedBy("TEST"); // 로그인 사용자 정보 들어가야 함
+        bookReport.getBookReportEntity().setMemberId(1L); // 로그인 사용자 정보 들어가야 함
+        bookReport.getBookReportEntity().setCreatedBy("TEST"); // 로그인 사용자 정보 들어가야 함
+        bookReportRepository.save(bookReport.getBookReportEntity());
+    }
+
+    public void deleteBookReport(Long bookReportId) {
+        bookReportRepository.findById(bookReportId)
+                .map(
+                        it -> {
+                            it.setIsDeleted("Y");
+                            it.setDeletedAt(LocalDateTime.now());
+                            bookReportRepository.save(it);
+                            return it;
+                        }
+                );
+    }
